Inline content from utils module
