+++ lib/link.js
@@ -1,7 +1,6 @@
 import React, { Component, Children } from 'react'
 import PropTypes from 'prop-types'
 import Router from '../client/router'
-import { warn, execOnce } from './utils'
 import { formatPath } from './url'
 
 export default class Link extends Component {
@@ -23,8 +22,8 @@ export default class Link extends Component {
       (props, propName) => {
         const value = props[propName]
 
-        if (typeof value === 'string') {
-          warnLink(`Warning: You're using a string directly inside <Link>. This usage has been deprecated. Please add an <a> tag as child of <Link>`)
+        if (typeof value === 'string' && process.env.NODE_ENV !== 'production') {
+          console.error(`Warning: You're using a string directly inside <Link>. This usage has been deprecated. Please add an <a> tag as child of <Link>`)
         }
 
         return null
@@ -106,5 +105,3 @@ export default class Link extends Component {
     return React.cloneElement(child, props)
   }
 }
-
-const warnLink = execOnce(warn)
