+++ lib/side-effect.js
@@ -1,7 +1,10 @@
 import React, { Component } from 'react'
-import { getDisplayName } from './utils'
 
 const canUseDOM = typeof window !== 'undefined'
+
+function getDisplayName (Component) {
+  return Component.displayName || Component.name || 'Unknown'
+}
 
 export default function withSideEffect (reduceComponentsToState, handleStateChangeOnClient) {
   return function wrap (WrappedComponent) {
