+++ lib/url.js
@@ -1,5 +1,11 @@
 import querystring from 'querystring'
 
+export function getURL () {
+  const { href, protocol, hostname, port } = window.location
+  const origin = `${protocol}//${hostname}${port ? ':' + port : ''}`
+  return href.substring(origin.length)
+}
+
 export function parsePath (url) {
   const hasHost = url.indexOf('//') >= 0
   url = url.replace(/^.*\/\//, '').replace(/#.*$/, '')
